rag_chatbot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
qdrant-client>=1.7.0
openai==1.3.5
pydantic==2.5.0
python-dotenv==1.0.0
asyncpg==0.29.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
numpy==1.24.3
tiktoken==0.5.1
langchain==0.0.339
langchain-openai==0.0.5
python-multipart==0.0.6